add authorizing condition

+++ app/Http/Controllers/UserController.php
@@ -15,8 +15,12 @@ class UserController extends Controller
      */
     public function index()
     {
-        $users = User::all();
-        return view('users')->with('users', $users);
+       if (AUTH::user()->hasRole('admin')) {
+            $users = User::all();
+            return view('users')->with('users', $users);
+     } else {
+            abort(401, 'This action is unauthorized.');
+        }
     }
 
     /**
@@ -27,12 +31,13 @@ class UserController extends Controller
      */
     public function destroy(Request $request, $id)
     {
-        /*TODO authorized Role wo festzulegen?*/
-        /*     $request->user()->authorizeRole('admin');*/
-        $user = User::find($id);
-        $user->delete();
-        Session::flash('alert-success', 'Benutzer ' . $user->email . ' wurde erfolgreich gelöscht.');
-        return redirect()->action('UserController@index');
-
+        if (AUTH::user()->hasRole('admin')) {
+            $user = User::find($id);
+            $user->delete();
+            Session::flash('alert-success', 'Benutzer ' . $user->email . ' wurde erfolgreich gelöscht.');
+            return redirect()->action('UserController@index');
+        } else {
+            abort(401, 'This action is unauthorized.');
+        }
     }
 }
